fu-hack-bot/requirements.txt
aiogram>=3.20.0.post0
langchain>=0.3.14
langchain-community>=0.3.19
langchain-openai==0.3.18
pandas>=2.2.3
openai>=1.82.0
python-dotenv>=1.0.1
Python>=3.12.7
google>=3.0.0
google-ai-generativelanguage>=0.6.15
google-api-core>=2.25.0rc1
google-api-python-client>=2.170.0
google-auth>=2.40.2
google-auth-httplib2>=0.2.0
google-genai>=1.16.1
google-generativeai>=0.8.5
google-pastav>=0.2.0
googleapis-common-protos>=1.70.0
  